HEEDLS-439 Reorganize data services and fix line endings
