+++ DigitalLearningSolutions.Web/Views/TrackingSystem/CourseSetup/AdminFields/RemoveAdminField.cshtml
@@ -11,7 +11,7 @@
   ViewData["HeaderPathName"] = "Tracking System";
   var confirmError = ViewData.ModelState[nameof(RemoveAdminFieldViewModel.Confirm)]?.Errors?.Count > 0;
   var confirmFormErrorClass = confirmError ? "nhsuk-form-group--error" : "";
-  var cancelLinkData = new Dictionary<string, string> { { "customisationId", Model.CustomisationId.ToString() } };
+  var cancelLinkRouteData = new Dictionary<string, string> { { "customisationId", Model.CustomisationId.ToString() } };
 }
 
 @section NavMenuItems {
@@ -74,6 +74,6 @@
       <button class="nhsuk-button delete-button nhsuk-button--secondary" type="submit">Delete field</button>
     </form>
 
-    <vc:cancel-link asp-controller="AdminFields" asp-action="AdminFields" asp-all-route-data="@cancelLinkData" />
+    <vc:cancel-link asp-controller="AdminFields" asp-action="AdminFields" asp-all-route-data="@cancelLinkRouteData" />
   </div>
 </div>
